==> TCGAdataProcessStep/2. frequency_on_tcga_data.py <==
# -*- coding: utf-8 -*-
"""frequency on TCGA data.ipynb

Automatically generated by Colaboratory.

Original file is located at
    https://colab.research.google.com/drive/1m1oICtGCZ5H4WRNT9e14KD0YlMpjESBz

# Imports
"""

import numpy as np
import pandas as pd
from sklearn.feature_selection import SelectKBest, chi2 , f_classif,f_regression,mutual_info_classif,mutual_info_regression
from sklearn.svm import SVR

"""# Frequency Check"""

cancer_names = ["BLCA","BRCA","CESC","CHOL","COAD","ESCA","HNSC","KICH","KIRC","KIRP","LIHC","LUAD","LUSC","PAAD","PCPG","PRAD","READ","SARC","STAD","THCA","THYM","UCEC"]

data = pd.read_csv("/home/vm21/PanClassify/cancer/CHOL.csv.gz",header=None)
one_col = data.iloc[:,0:1]
one_col = one_col.drop(one_col.index[0])

gene_frequency = []
#reading data and doing work
for index in range(len(cancer_names)):
  DATA = np.load("/home/vm21/PanClassify/std_npy/"+cancer_names[index]+".npy")
  for i in range(len(DATA)):
    if (DATA[i] != 0):
      gene_frequency.append(one_col[0][DATA[i]])

np.save("/home/vm21/PanClassify/std_npy/gene_frequency",gene_frequency)
